Feat: Redesign Dashboard UI to Match New Design
This commit completely overhauls the dashboard UI to align with the new design specifications provided in the reference screenshot.

Key changes include:
- Restructured the entire dashboard layout using CSS Grid for a more robust and responsive design.
- Updated the sidebar navigation with new items, icons, and a user profile section.
- Reordered and restyled the top metric cards with new background colors, icons, and typography.
- Replaced the previous main content area with static SVG representations of the "AI Success Rate Over Time" line chart and the "Denial Types Distribution" pie chart.
- Added a new bottom section with "Recent Activity" and "AI Learning Insights" panels.
- Refreshed the color palette, typography, and spacing to create a modern and clean user interface.
- Temporarily commented out the company logo to allow the application to build and render, as the asset was missing from the project.
