+++ Kuve-AKU-1/npm_output.log
@@ -2,8 +2,9 @@
 > vite-react-app@0.0.0 dev
 > vite
 
+Port 5173 is in use, trying another one...
 
-  VITE v5.4.20  ready in 441 ms
+  VITE v5.4.20  ready in 797 ms
 
-  ➜  Local:   http://localhost:5173/
+  ➜  Local:   http://localhost:5174/
   ➜  Network: use --host to expose
