+++ Kuve-AKU-1/vite.config.ts
@@ -4,6 +4,7 @@ import path from 'path'
 
 // https://vitejs.dev/config/
 export default defineConfig({
+  publicDir: 'public',
   plugins: [react()],
   resolve: {
     alias: {
