+++ npm_output.log
@@ -2,13 +2,8 @@
 > vite-react-app@0.0.0 dev
 > vite
 
-Port 5173 is in use, trying another one...
-Port 5174 is in use, trying another one...
-Port 5175 is in use, trying another one...
-Port 5176 is in use, trying another one...
-Port 5177 is in use, trying another one...
 
-  VITE v5.4.20  ready in 784 ms
+  VITE v5.4.20  ready in 369 ms
 
-  ➜  Local:   http://localhost:5178/
+  ➜  Local:   http://localhost:5173/
   ➜  Network: use --host to expose
